Break up character_functions.cpp
