+++ src/avatar_action.cpp
@@ -17,8 +17,8 @@
 #include "bodypart.h"
 #include "calendar.h"
 #include "character.h"
-#include "character_functions.h"
 #include "character_martial_arts.h"
+#include "character_turn.h"
 #include "creature.h"
 #include "cursesdef.h"
 #include "debug.h"
