+++ src/character_functions.cpp
@@ -1,5 +1,4 @@
 #include "character_functions.h"
-#include "character_effects.h"
 
 #include "bionics.h"
 #include "calendar.h"
@@ -404,76 +403,6 @@ bool roll_can_sleep( Character &who )
     return result;
 }
 
-void update_body_wetness( Character &who, const w_point &weather )
-{
-    // Average number of turns to go from completely soaked to fully dry
-    // assuming average temperature and humidity
-    constexpr time_duration average_drying = 2_hours;
-
-    // A modifier on drying time
-    double delay = 1.0;
-    // Weather slows down drying
-    delay += ( ( weather.humidity - 66 ) - ( units::to_fahrenheit( weather.temperature ) - 65 ) ) / 100;
-    delay = std::max( 0.1, delay );
-    // Fur/slime retains moisture
-    if( who.has_trait( trait_LIGHTFUR ) ||
-        who.has_trait( trait_FUR ) ||
-        who.has_trait( trait_FELINE_FUR ) ||
-        who.has_trait( trait_LUPINE_FUR ) ||
-        who.has_trait( trait_CHITIN_FUR ) ||
-        who.has_trait( trait_CHITIN_FUR2 ) ||
-        who.has_trait( trait_CHITIN_FUR3 ) ) {
-        delay = delay * 6 / 5;
-    }
-    if( who.has_trait( trait_URSINE_FUR ) || who.has_trait( trait_SLIMY ) ) {
-        delay *= 1.5;
-    }
-
-    if( !x_in_y( 1, to_turns<int>( average_drying * delay / 100.0 ) ) ) {
-        // No drying this turn
-        return;
-    }
-
-    // Now per-body-part stuff
-    // To make drying uniform, make just one roll and reuse it
-    const int drying_roll = rng( 1, 80 );
-
-    for( const body_part bp : all_body_parts ) {
-        if( who.body_wetness[bp] == 0 ) {
-            continue;
-        }
-        // This is to normalize drying times
-        int drying_chance = who.drench_capacity[bp];
-        // Body temperature affects duration of wetness
-        // Note: Using temp_conv rather than temp_cur, to better approximate environment
-        if( who.temp_conv[bp] >= BODYTEMP_SCORCHING ) {
-            drying_chance *= 2;
-        } else if( who.temp_conv[bp] >= BODYTEMP_VERY_HOT ) {
-            drying_chance = drying_chance * 3 / 2;
-        } else if( who.temp_conv[bp] >= BODYTEMP_HOT ) {
-            drying_chance = drying_chance * 4 / 3;
-        } else if( who.temp_conv[bp] > BODYTEMP_COLD ) {
-            // Comfortable, doesn't need any changes
-        } else {
-            // Evaporation doesn't change that much at lower temp
-            drying_chance = drying_chance * 3 / 4;
-        }
-
-        if( drying_chance < 1 ) {
-            drying_chance = 1;
-        }
-
-        // TODO: Make evaporation reduce body heat
-        if( drying_chance >= drying_roll ) {
-            who.body_wetness[bp] -= 1;
-            if( who.body_wetness[bp] < 0 ) {
-                who.body_wetness[bp] = 0;
-            }
-        }
-    }
-    // TODO: Make clothing slow down drying
-}
-
 bool can_interface_armor( const Character &who )
 {
     bool okay = std::any_of( who.my_bionics->begin(), who.my_bionics->end(),
@@ -483,283 +412,4 @@ bool can_interface_armor( const Character &who )
     return okay;
 }
 
-void do_pause( Character &who )
-{
-    map &here = get_map();
-
-    who.moves = 0;
-    who.recoil = MAX_RECOIL;
-
-    // Train swimming if underwater
-    if( !who.in_vehicle ) {
-        if( who.is_underwater() ) {
-            who.as_player()->practice( skill_swimming, 1 );
-            who.drench( 100, { {
-                    bp_leg_l, bp_leg_r, bp_torso, bp_arm_l,
-                    bp_arm_r, bp_head, bp_eyes, bp_mouth,
-                    bp_foot_l, bp_foot_r, bp_hand_l, bp_hand_r
-                }
-            }, true );
-        } else if( here.has_flag( TFLAG_DEEP_WATER, who.pos() ) ) {
-            who.as_player()->practice( skill_swimming, 1 );
-            // Same as above, except no head/eyes/mouth
-            who.drench( 100, { {
-                    bp_leg_l, bp_leg_r, bp_torso, bp_arm_l,
-                    bp_arm_r, bp_foot_l, bp_foot_r, bp_hand_l,
-                    bp_hand_r
-                }
-            }, true );
-        } else if( here.has_flag( "SWIMMABLE", who.pos() ) ) {
-            who.drench( 40, { { bp_foot_l, bp_foot_r, bp_leg_l, bp_leg_r } }, false );
-        }
-    }
-
-    // Try to put out clothing/hair fire
-    if( who.has_effect( effect_onfire ) ) {
-        time_duration total_removed = 0_turns;
-        time_duration total_left = 0_turns;
-        bool on_ground = who.has_effect( effect_downed );
-        for( const body_part bp : all_body_parts ) {
-            effect &eff = who.get_effect( effect_onfire, bp );
-            if( eff.is_null() ) {
-                continue;
-            }
-
-            // TODO: Tools and skills
-            total_left += eff.get_duration();
-            // Being on the ground will smother the fire much faster because you can roll
-            const time_duration dur_removed = on_ground ? eff.get_duration() / 2 + 2_turns : 1_turns;
-            eff.mod_duration( -dur_removed );
-            total_removed += dur_removed;
-        }
-
-        // Don't drop on the ground when the ground is on fire
-        if( total_left > 1_minutes && !who.is_dangerous_fields( here.field_at( who.pos() ) ) ) {
-            who.add_effect( effect_downed, 2_turns, num_bp, 0, true );
-            who.add_msg_player_or_npc( m_warning,
-                                       _( "You roll on the ground, trying to smother the fire!" ),
-                                       _( "<npcname> rolls on the ground!" ) );
-        } else if( total_removed > 0_turns ) {
-            who.add_msg_player_or_npc( m_warning,
-                                       _( "You attempt to put out the fire on you!" ),
-                                       _( "<npcname> attempts to put out the fire on them!" ) );
-        }
-    }
-
-    // on-pause effects for martial arts
-    who.martial_arts_data->ma_onpause_effects( who );
-
-    if( who.is_npc() ) {
-        // The stuff below doesn't apply to NPCs
-        // search_surroundings should eventually do, though
-        return;
-    }
-
-    if( who.in_vehicle && one_in( 8 ) ) {
-        VehicleList vehs = here.get_vehicles();
-        vehicle *veh = nullptr;
-        for( auto &v : vehs ) {
-            veh = v.v;
-            if( veh && veh->is_moving() && veh->player_in_control( who ) ) {
-                double exp_temp = 1 + veh->total_mass() / 400.0_kilogram +
-                                  std::abs( veh->velocity / 3200.0 );
-                int experience = static_cast<int>( exp_temp );
-                if( exp_temp - experience > 0 && x_in_y( exp_temp - experience, 1.0 ) ) {
-                    experience++;
-                }
-                who.as_player()->practice( skill_id( "driving" ), experience );
-                break;
-            }
-        }
-    }
-
-    search_surroundings( who );
-    who.wait_effects();
-}
-
-void search_surroundings( Character &who )
-{
-    if( who.controlling_vehicle ) {
-        return;
-    }
-    const map &here = get_map();
-    // Search for traps in a larger area than before because this is the only
-    // way we can "find" traps that aren't marked as visible.
-    // Detection formula takes care of likelihood of seeing within this range.
-    for( const tripoint &tp : here.points_in_radius( who.pos(), 5 ) ) {
-        const trap &tr = here.tr_at( tp );
-        if( tr.is_null() || tp == who.pos() ) {
-            continue;
-        }
-        if( who.has_active_bionic( bio_ground_sonar ) && !who.knows_trap( tp ) &&
-            ( tr.loadid == tr_beartrap_buried ||
-              tr.loadid == tr_landmine_buried || tr.loadid == tr_sinkhole ) ) {
-            const std::string direction = direction_name( direction_from( who.pos(), tp ) );
-            who.add_msg_if_player( m_warning, _( "Your ground sonar detected a %1$s to the %2$s!" ),
-                                   tr.name(), direction );
-            who.add_known_trap( tp, tr );
-        }
-        if( !who.sees( tp ) ) {
-            continue;
-        }
-        if( tr.is_always_invisible() || tr.can_see( tp, who ) ) {
-            // Already seen, or can never be seen
-            continue;
-        }
-        // Chance to detect traps we haven't yet seen.
-        if( tr.detect_trap( tp, who ) ) {
-            if( tr.get_visibility() > 0 ) {
-                // Only bug player about traps that aren't trivial to spot.
-                const std::string direction = direction_name(
-                                                  direction_from( who.pos(), tp ) );
-                who.add_msg_if_player( _( "You've spotted a %1$s to the %2$s!" ),
-                                       tr.name(), direction );
-            }
-            who.add_known_trap( tp, tr );
-        }
-    }
-}
-
 } // namespace character_funcs
-
-namespace character_effects
-{
-
-stat_mod get_pain_penalty( const Character &ch )
-{
-    stat_mod ret;
-    int pain = ch.get_perceived_pain();
-    if( pain <= 0 ) {
-        return ret;
-    }
-
-    int stat_penalty = std::floor( std::pow( pain, 0.8f ) / 10.0f );
-
-    bool ceno = ch.has_trait( trait_CENOBITE );
-    if( !ceno ) {
-        ret.strength = stat_penalty;
-        ret.dexterity = stat_penalty;
-    }
-
-    if( !ch.has_trait( trait_INT_SLIME ) ) {
-        ret.intelligence = stat_penalty;
-    } else {
-        ret.intelligence = pain / 5;
-    }
-
-    ret.perception = stat_penalty * 2 / 3;
-
-    ret.speed = std::pow( pain, 0.7f );
-    if( ceno ) {
-        ret.speed /= 2;
-    }
-
-    ret.speed = std::min( ret.speed, 30 );
-    return ret;
-}
-
-int get_kcal_speed_penalty( float kcal_percent )
-{
-    static const std::vector<std::pair<float, float>> starv_thresholds = { {
-            std::make_pair( 0.0f, -90.0f ),
-            std::make_pair( 0.1f, -50.f ),
-            std::make_pair( 0.3f, -25.0f ),
-            std::make_pair( 0.5f, 0.0f )
-        }
-    };
-    if( kcal_percent > 0.95f ) {
-        return 0;
-    } else {
-        return std::round( multi_lerp( starv_thresholds, kcal_percent ) );
-    }
-}
-
-int get_thirst_speed_penalty( int thirst )
-{
-    // We die at 1200 thirst
-    // Start by dropping speed really fast, but then level it off a bit
-    static const std::vector<std::pair<float, float>> thirst_thresholds = {{
-            std::make_pair( static_cast<float>( thirst_levels::very_thirsty ), 0.0f ),
-            std::make_pair( static_cast<float>( thirst_levels::dehydrated ), -25.0f ),
-            std::make_pair( static_cast<float>( thirst_levels::parched ), -50.0f ),
-            std::make_pair( static_cast<float>( thirst_levels::dead ), -75.0f )
-        }
-    };
-    return static_cast<int>( multi_lerp( thirst_thresholds, thirst ) );
-}
-
-int calc_morale_fatigue_cap( int fatigue )
-{
-    if( fatigue >= fatigue_levels::massive ) {
-        return 20;
-    } else if( fatigue >= fatigue_levels::exhausted ) {
-        return 40;
-    } else if( fatigue >= fatigue_levels::dead_tired ) {
-        return 60;
-    } else if( fatigue >= fatigue_levels::tired ) {
-        return 80;
-    }
-    return 0;
-}
-
-double vomit_mod( const Character &ch )
-{
-    double mod = 1;
-    if( ch.has_effect( effect_weed_high ) ) {
-        mod *= .1;
-    }
-    if( ch.has_trait( trait_STRONGSTOMACH ) ) {
-        mod *= .5;
-    }
-    if( ch.has_trait( trait_WEAKSTOMACH ) ) {
-        mod *= 2;
-    }
-    if( ch.has_trait( trait_NAUSEA ) ) {
-        mod *= 3;
-    }
-    if( ch.has_trait( trait_VOMITOUS ) ) {
-        mod *= 3;
-    }
-    // If you're already nauseous, any food in your stomach greatly
-    // increases chance of vomiting. Water doesn't provoke vomiting, though.
-    if( ch.stomach.get_calories() > 0 && ch.has_effect( effect_nausea ) ) {
-        mod *= 5 * ch.get_effect_int( effect_nausea );
-    }
-    return mod;
-}
-
-int talk_skill( const Character &ch )
-{
-    /** @EFFECT_INT slightly increases talking skill */
-
-    /** @EFFECT_PER slightly increases talking skill */
-
-    /** @EFFECT_SPEECH increases talking skill */
-    int ret = ch.get_int() + ch.get_per() + ch.get_skill_level( skill_id( "speech" ) ) * 3;
-    return ret;
-}
-
-int intimidation( const Character &ch )
-{
-    /** @EFFECT_STR increases intimidation factor */
-    int ret = ch.get_str() * 2;
-    if( ch.weapon.is_gun() ) {
-        ret += 10;
-    }
-    if( ch.weapon.damage_melee( DT_BASH ) >= 12 ||
-        ch.weapon.damage_melee( DT_CUT ) >= 12 ||
-        ch.weapon.damage_melee( DT_STAB ) >= 12 ) {
-        ret += 5;
-    }
-
-    if( ch.get_stim() > 20 ) {
-        ret += 2;
-    }
-    if( ch.has_effect( effect_drunk ) ) {
-        ret -= 4;
-    }
-
-    return ret;
-}
-
-} // namespace character_effects
