+++ src/character_functions.h
@@ -10,7 +10,6 @@ class item;
 class time_duration;
 class vehicle;
 struct tripoint;
-struct w_point;
 
 namespace character_funcs
 {
@@ -95,17 +94,8 @@ int rate_sleep_spot( const Character &who, const tripoint &p );
 /** Checked each turn during "lying_down", returns true if the avatar falls asleep */
 bool roll_can_sleep( Character &who );
 
-/** Maintains body wetness and handles the rate at which the player dries */
-void update_body_wetness( Character &who, const w_point &weather );
-
 /** Check whether character has an active bionic capable of interfacing with power armor. */
 bool can_interface_armor( const Character &who );
-
-/** Do pause action ('.' key). */
-void do_pause( Character &who );
-
-/** Search surrounding squares for traps (and maybe other things in the future). */
-void search_surroundings( Character &who );
 
 } // namespace character_funcs
