+++ src/handle_action.cpp
@@ -21,8 +21,8 @@
 #include "catacharset.h"
 #include "character.h"
 #include "character_display.h"
-#include "character_functions.h"
 #include "character_martial_arts.h"
+#include "character_turn.h"
 #include "clzones.h"
 #include "color.h"
 #include "construction.h"
